refactor: move collapsible-panel
