+++ packages/components/collapsible-panel/README.md
