+++ packages/components/collapsible-panel/src/collapsible-panel-header.js
