+++ packages/components/collapsible-panel/src/collapsible-panel.js
