+++ packages/components/collapsible-panel/src/collapsible-panel.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render } from '../../../test-utils';
+import { render } from '../../../../src/test-utils';
 import CollapsiblePanel from './collapsible-panel';
 
 const HeaderControls = props => (
