+++ packages/components/collapsible-panel/src/collapsible-panel.story.js
@@ -4,7 +4,7 @@ import { withKnobs, text, boolean, select } from '@storybook/addon-knobs/react';
 import Section from '../../../../.storybook/decorators/section';
 import CollapsiblePanel from './collapsible-panel';
 import CollapsiblePanelHeader from './collapsible-panel-header';
-import Readme from './README.md';
+import Readme from '../README.md';
 
 storiesOf('Components|Panels', module)
   .addDecorator(withKnobs)
