+++ packages/components/collapsible-panel/src/collapsible-panel.styles.js
