+++ packages/components/collapsible-panel/src/collapsible-panel.visualroute.js
