+++ packages/components/collapsible-panel/src/collapsible-panel.visualspec.js
