+++ packages/components/collapsible-panel/src/header-icon.js
