+++ packages/components/collapsible-panel/src/index.js
